+++ src/hooks/useMiniStudiesQuery.ts
